+++ ext/standard/basic_functions.c
@@ -305,7 +305,6 @@ PHP_MINIT_FUNCTION(basic) /* {{{ */
 	BASIC_MINIT_SUBMODULE(user_filters)
 	BASIC_MINIT_SUBMODULE(password)
 	BASIC_MINIT_SUBMODULE(image)
-	BASIC_MINIT_SUBMODULE(url)
 
 #ifdef ZTS
 	BASIC_MINIT_SUBMODULE(localeconv)
