+++ ext/standard/url.c
@@ -25,8 +25,6 @@
 #include "file.h"
 #include "zend_simd.h"
 #include "Zend/zend_smart_str.h"
-#include "Zend/zend_exceptions.h"
-#include "ext/uri/php_uri.h"
 
 /* {{{ free_url */
 PHPAPI void php_url_free(php_url *theurl)
@@ -48,13 +46,6 @@ PHPAPI void php_url_free(php_url *theurl)
 	efree(theurl);
 }
 /* }}} */
-
-static void parse_url_free_uri(void *uri)
-{
-	php_url *parse_url_uri = (php_url *) uri;
-
-	php_url_free(parse_url_uri);
-}
 
 static void php_replace_controlchars(char *str, size_t len)
 {
@@ -320,166 +311,7 @@ parse_host:
 
 	return ret;
 }
-
-static void parse_url_decode_component(zval *zv, uri_component_read_mode_t read_mode)
-{
-	if (Z_TYPE_P(zv) != IS_STRING) {
-		return;
-	}
-
-	if (read_mode == URI_COMPONENT_READ_RAW) {
-		return;
-	}
-
-	php_raw_url_decode(Z_STRVAL_P(zv), Z_STRLEN_P(zv));
-}
-
-static zend_result parse_url_read_scheme(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->scheme) {
-		ZVAL_STR_COPY(retval, parse_url_uri->scheme);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static zend_result parse_url_read_username(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->user) {
-		ZVAL_STR_COPY(retval, parse_url_uri->user);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static zend_result parse_url_read_password(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->pass) {
-		ZVAL_STR_COPY(retval, parse_url_uri->pass);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static zend_result parse_url_read_host(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->host) {
-		ZVAL_STR_COPY(retval, parse_url_uri->host);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static zend_result parse_url_read_port(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->port) {
-		ZVAL_LONG(retval, parse_url_uri->port);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static zend_result parse_url_read_path(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->path) {
-		ZVAL_STR_COPY(retval, parse_url_uri->path);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static zend_result parse_url_read_query(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->query) {
-		ZVAL_STR_COPY(retval, parse_url_uri->query);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static zend_result parse_url_read_fragment(const uri_internal_t *internal_uri, uri_component_read_mode_t read_mode, zval *retval)
-{
-	php_url *parse_url_uri = internal_uri->uri;
-
-	if (parse_url_uri->fragment) {
-		ZVAL_STR_COPY(retval, parse_url_uri->fragment);
-		parse_url_decode_component(retval, read_mode);
-	} else {
-		ZVAL_NULL(retval);
-	}
-
-	return SUCCESS;
-}
-
-static void throw_invalid_uri_exception(void)
-{
-	zend_throw_exception(uri_invalid_uri_exception_ce, "The specified URI is malformed", 0);
-}
-
-static void *parse_url_parse_uri(const char *uri_str, size_t uri_str_len, const void *base_url, zval *errors, bool silent)
-{
-	bool has_port;
-
-	php_url *url = php_url_parse_ex2(uri_str, uri_str_len, &has_port);
-	if (url == NULL && !silent) {
-		throw_invalid_uri_exception();
-	}
-
-	return url;
-}
-
-const uri_parser_t parse_url_uri_parser = {
-	.name = URI_PARSER_PHP,
-	.parse_uri = parse_url_parse_uri,
-	.clone_uri = NULL,
-	.uri_to_string = NULL,
-	.free_uri = parse_url_free_uri,
-	{
-		.scheme = {.read_func = parse_url_read_scheme, .write_func = NULL},
-		.username = {.read_func = parse_url_read_username, .write_func = NULL},
-		.password = {.read_func = parse_url_read_password, .write_func = NULL},
-		.host = {.read_func = parse_url_read_host, .write_func = NULL},
-		.port = {.read_func = parse_url_read_port, .write_func = NULL},
-		.path = {.read_func = parse_url_read_path, .write_func = NULL},
-		.query = {.read_func = parse_url_read_query, .write_func = NULL},
-		.fragment = {.read_func = parse_url_read_fragment, .write_func = NULL},
-	}
-};
+/* }}} */
 
 /* {{{ Parse a URL and return its components */
 PHP_FUNCTION(parse_url)
@@ -921,8 +753,3 @@ no_name_header:
 	php_stream_close(stream);
 }
 /* }}} */
-
-PHP_MINIT_FUNCTION(url)
-{
-	return php_uri_parser_register(&parse_url_uri_parser);
-}
